requirements.txt
requests
pandas
openpyxl
ttkbootstrap
python-dotenv==1.0.1